requirements.txt
anthropic>=0.25.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
